refactor: map - populate algolia scripts
