+++ map/src/algolia-scripts/config-algolia-index.js
