+++ map/src/algolia-scripts/export-algolia-index.js
